likeボタン

+++ app/Http/Controllers/PostController.php
@@ -7,6 +7,9 @@ use App\Post;
 use Illuminate\Http\Request;
 use App\Category;
 use App\Comment;
+use App\User;
+use App\Like;
+use Illuminate\Support\Facades\Auth;
 
 class PostController extends Controller
 {
@@ -15,11 +18,12 @@ class PostController extends Controller
        return view('posts/index')->with(['posts' => $post->getPaginateByLimit()]);
    }
  
-   public function show(Post $post)
+   public function show(Post $post, User $user)
    {
       return view('posts/show')->with([
-        'post' => $post,
-        'comments' => $post->getCommentsPaginate(),
+       'post' => $post,
+       'comments' => $post->getCommentsPaginate(),
+       'user' => $user,
       ]);
    }
 
@@ -31,6 +35,7 @@ class PostController extends Controller
    public function store(PostRequest $request, Post $post)
    {
        $input = $request['post'];
+       $input['user_id'] = Auth::id();
        $created_post = $post->createWithRelation($input);
        return redirect('/posts/' . $created_post->id);
    }
@@ -46,6 +51,7 @@ class PostController extends Controller
     
    public function update(PostRequest $request, Post $post)
    {
+       $this->authorize('update',$post);
        $input = $request['post'];
        $target_post = $post->updateWithRelation($input);
        return redirect('/posts/' . $target_post->id);
@@ -53,7 +59,39 @@ class PostController extends Controller
    
    public function destroy(Post $post)
    {
+      $this->authorize('delete',$post);
       $post->deleteWithRelation();
       return redirect('/');
+   }
+   
+   // only()の引数内のメソッドはログイン時のみ有効
+   // public function __construct()
+   // {
+   //    $this->middleware(['auth', 'verified'])->only(['like', 'unlike']);
+   // }
+   public function like($id)
+   {
+      Like::create([
+         'post_id' => $id,
+         'user_id' => Auth::id(),
+      ]);
+      
+      session()->flash('success','You Liked the Post');
+      
+      return redirect()->back();
+   }
+   
+   public function unlike($id)
+   {
+      $like = Like::where('post_id',$id)->where('user_id',Auth::id())->first();
+      $like->delete();
+      
+      session()->flash('success', 'You Unliked the Post.');
+      
+      return redirect()->back();
+   }
+   
+   public function hello() {
+      return 'hello';
    }
 }
